+++ packages/gatsby-plugin-twitter/src/gatsby-browser.js
@@ -10,27 +10,34 @@ const injectTwitterScript = () => {
     window.twttr = (function(d, s, id) {
       var js,
         fjs = d.getElementsByTagName(s)[0],
-        t = window.twttr || {}
-      if (d.getElementById(id)) return t
-      js = d.createElement(s)
-      js.id = id
-      js.src = "https://platform.twitter.com/widgets.js"
-      fjs.parentNode.insertBefore(js, fjs)
-      t._e = []
+        t = window.twttr || {};
+      if (d.getElementById(id)) return t;
+      js = d.createElement(s);
+      js.id = id;
+      js.src = "https://platform.twitter.com/widgets.js";
+      fjs.parentNode.insertBefore(js, fjs);
+      t._e = [];
       t.ready = function(f) {
-        t._e.push(f)
-      }
-      return t
-    })(document, "script", "twitter-wjs")
+        t._e.push(f);
+      };
+      return t;
+    })(document, "script", "twitter-wjs");
   `)
 }
 
 let injectedTwitterScript = false
 
+const embedClasses = [
+  `.twitter-tweet`,
+  `.twitter-timeline`,
+  `.twitter-follow-button`,
+  `.twitter-share-button`,
+].join(`,`)
+
 exports.onRouteUpdate = () => {
-  // If there's an embedded tweet, lazy-load the twitter script (if it hasn't
+  // If there's an embedded element, lazy-load the twitter script (if it hasn't
   // already been loaded), and then run the twitter load function.
-  if (document.querySelector(`.twitter-tweet`) !== null) {
+  if (document.querySelector(embedClasses) !== null) {
     if (!injectedTwitterScript) {
       injectTwitterScript()
       injectedTwitterScript = true