+++ packages/gatsby-remark-prismjs/src/__tests__/highlight-code.js
@@ -56,11 +56,42 @@ export default Counter
   })
 
   describe(`with language-text`, () => {
-    it(`escapes &, <, " elements #4597`, () => {
+    it(`escapes &, <, " elements #4597 and warns`, () => {
+      spyOn(console, `warn`)
+
       const highlightCode = require(`../highlight-code`)
       const language = `text`
       const code = `<button />`
       expect(highlightCode(language, code)).toMatch(`&lt;button /&gt;`)
+      expect(console.warn).toHaveBeenCalledWith(
+        `unable to find prism language 'text' for highlighting.`,
+        `applying generic code block`
+      )
+    })
+
+    it(`warns once per language`, () => {
+      spyOn(console, `warn`)
+
+      const highlightCode = require(`../highlight-code`)
+      const language1 = `text`
+      const language2 = `raw`
+      const code1 = `<button />`
+      const code2 = `<form />`
+      const code3 = `<input />`
+      highlightCode(language1, code1)
+      highlightCode(language1, code2)
+      highlightCode(language2, code3)
+      expect(console.warn).toHaveBeenCalledTimes(2)
+      expect(console.warn).toHaveBeenNthCalledWith(
+        1,
+        `unable to find prism language 'text' for highlighting.`,
+        `applying generic code block`
+      )
+      expect(console.warn).toHaveBeenNthCalledWith(
+        2,
+        `unable to find prism language 'raw' for highlighting.`,
+        `applying generic code block`
+      )
     })
   })