+++ packages/gatsby/cache-dir/__tests__/static-entry.js
@@ -17,7 +17,9 @@ jest.mock(
       },
     }
   },
-  { virtual: true }
+  {
+    virtual: true,
+  }
 )
 
 jest.mock(
@@ -38,7 +40,9 @@ jest.mock(
       ],
     }
   },
-  { virtual: true }
+  {
+    virtual: true,
+  }
 )
 
 const MOCK_FILE_INFO = {
@@ -61,13 +65,48 @@ const reverseHeadersPlugin = {
   },
 }
 
+const injectValuePlugin = (hookName, methodName, value) => {
+  return {
+    plugin: {
+      [hookName]: staticEntry => {
+        const method = staticEntry[methodName]
+        method(value)
+      },
+    },
+  }
+}
+
+const checkSanitized = components => {
+  expect(components.includes(null)).toBeFalsy()
+  expect(
+    components.find(val => Array.isArray(val) && val.length === 0)
+  ).toBeFalsy()
+}
+
+const checkNonEmptyHeadersPlugin = {
+  plugin: {
+    onPreRenderHTML: ({
+      getHeadComponents,
+      getPreBodyComponents,
+      getPostBodyComponents,
+    }) => {
+      const headComponents = getHeadComponents()
+      const preBodyComponents = getPreBodyComponents()
+      const postBodyComponents = getPostBodyComponents()
+      checkSanitized(headComponents)
+      checkSanitized(preBodyComponents)
+      checkSanitized(postBodyComponents)
+    },
+  },
+}
+
 const fakeStylesPlugin = {
   plugin: {
     onRenderBody: ({ setHeadComponents }) =>
       setHeadComponents([
-        <style key="style1">.style1 {}</style>,
-        <style key="style2">.style2 {}</style>,
-        <style key="style3">.style3 {}</style>,
+        <style key="style1"> .style1 {} </style>,
+        <style key="style2"> .style2 {} </style>,
+        <style key="style3"> .style3 {} </style>,
       ]),
   },
 }
@@ -89,9 +128,9 @@ const fakeComponentsPluginFactory = type => {
     plugin: {
       onRenderBody: props => {
         props[`set${type}BodyComponents`]([
-          <div key="div1">div1</div>,
-          <div key="div2">div2</div>,
-          <div key="div3">div3</div>,
+          <div key="div1"> div1 </div>,
+          <div key="div2"> div2 </div>,
+          <div key="div3"> div3 </div>,
         ])
       },
     },
@@ -129,6 +168,59 @@ describe(`develop-static-entry`, () => {
     DevelopStaticEntry(`/about/`, (_, html) => {
       expect(html).toMatchSnapshot()
       done()
+    })
+  })
+})
+
+describe(`static-entry sanity checks`, () => {
+  beforeEach(() => {
+    global.__PATH_PREFIX__ = ``
+  })
+
+  const methodsToCheck = [
+    `replaceHeadComponents`,
+    `replacePreBodyComponents`,
+    `replacePostBodyComponents`,
+  ]
+
+  methodsToCheck.forEach(methodName => {
+    test(`${methodName} can filter out null value`, done => {
+      const plugin = injectValuePlugin(`onPreRenderHTML`, methodName, null)
+      global.plugins = [plugin, checkNonEmptyHeadersPlugin]
+
+      StaticEntry(`/about/`, (_, html) => {
+        done()
+      })
+    })
+
+    test(`${methodName} can filter out null values`, done => {
+      const plugin = injectValuePlugin(`onPreRenderHTML`, methodName, [
+        null,
+        null,
+      ])
+      global.plugins = [plugin, checkNonEmptyHeadersPlugin]
+
+      StaticEntry(`/about/`, (_, html) => {
+        done()
+      })
+    })
+
+    test(`${methodName} can filter out empty array`, done => {
+      const plugin = injectValuePlugin(`onPreRenderHTML`, methodName, [])
+      global.plugins = [plugin, checkNonEmptyHeadersPlugin]
+
+      StaticEntry(`/about/`, (_, html) => {
+        done()
+      })
+    })
+
+    test(`${methodName} can filter out empty arrays`, done => {
+      const plugin = injectValuePlugin(`onPreRenderHTML`, methodName, [[], []])
+      global.plugins = [plugin, checkNonEmptyHeadersPlugin]
+
+      StaticEntry(`/about/`, (_, html) => {
+        done()
+      })
     })
   })
 })