+++ www/src/components/ecosystem/ecosystem-board.js
@@ -4,7 +4,7 @@ import styled from "@emotion/styled"
 
 import EcosystemSection from "./ecosystem-section"
 
-import presets from "../../utils/presets"
+import { space, breakpoints, dimensions } from "../../utils/presets"
 import {
   setupScrollersObserver,
   unobserveScrollers,
@@ -14,13 +14,13 @@ const EcosystemBoardRoot = styled(`div`)`
   display: flex;
   flex-direction: column;
 
-  ${presets.Md} {
+  ${breakpoints.md} {
     flex-direction: row;
     flex-wrap: wrap;
     height: calc(
-      100vh - (${presets.bannerHeight} + ${presets.headerHeight} + 1px)
+      100vh - (${dimensions.bannerHeight} + ${dimensions.headerHeight} + 1px)
     );
-    padding: 2rem 1rem 1rem;
+    padding: ${space[7]} ${space[4]} ${space[4]};
   }
 `