+++ www/src/components/navigation.js
@@ -5,13 +5,21 @@ import TwitterIcon from "react-icons/lib/fa/twitter"
 import SearchForm from "../components/search-form"
 import DiscordIcon from "../components/discord"
 import logo from "../logo.svg"
-import { rhythm, scale, options } from "../utils/typography"
-import presets, { colors } from "../utils/presets"
+import { rhythm } from "../utils/typography"
+import {
+  colors,
+  space,
+  scale,
+  transition,
+  breakpoints,
+  dimensions,
+  fonts,
+} from "../utils/presets"
 
 // what we need to nudge down the navItems to sit
 // on the baseline of the logo's wordmark
-const navItemTopOffset = `0.6rem`
-const navItemHorizontalSpacing = rhythm(1 / 3)
+const navItemTopOffset = `0.4rem`
+const navItemHorizontalSpacing = space[2]
 
 const iconColor = colors.lilac
 
@@ -41,6 +49,7 @@ const Navigation = ({ pathname }) => {
       title={title}
       css={{
         ...socialIconsStyles,
+        fontSize: scale[2],
         ...overrideCSS,
       }}
     >
@@ -53,22 +62,11 @@ const Navigation = ({ pathname }) => {
       css={{
         backgroundColor: `rgba(255,255,255,0.975)`,
         position: `relative`,
-        height: presets.headerHeight,
+        height: dimensions.headerHeight,
         left: 0,
         right: 0,
-        top: presets.bannerHeight,
+        top: dimensions.bannerHeight,
         zIndex: 2,
-        "&:after": {
-          content: `''`,
-          position: `absolute`,
-          bottom: 0,
-          left: 0,
-          right: 0,
-          width: `100%`,
-          height: 1,
-          zIndex: -1,
-          background: isHomepage ? `transparent` : colors.ui.light,
-        },
         // use this to test if the header items are properly aligned to the logo
         // wordmark
         // "&:before": {
@@ -82,15 +80,29 @@ const Navigation = ({ pathname }) => {
         //   zIndex: 10,
         //   background: `red`,
         // },
-        [presets.Md]: {
+        [breakpoints.md]: {
           position: isHomepage || isBlog ? `absolute` : `fixed`,
-          backgroundColor: isBlog ? colors.ui.whisper : false,
+          backgroundColor: isBlog ? colors.gray.whisper : false,
         },
         paddingLeft: `env(safe-area-inset-left)`,
         paddingRight: `env(safe-area-inset-right)`,
       }}
     >
-      <div css={{ ...styles.containerInner }}>
+      <div
+        css={{
+          ...styles.containerInner,
+          "&:after": {
+            content: `''`,
+            position: `absolute`,
+            bottom: 0,
+            left: 0,
+            right: 0,
+            height: 1,
+            zIndex: -1,
+            background: isHomepage ? `transparent` : colors.gray.border,
+          },
+        }}
+      >
         <Link
           to="/"
           css={styles.logoLink}
@@ -142,8 +154,7 @@ const Navigation = ({ pathname }) => {
           <div
             css={{
               display: `none`,
-              [presets.Lg]: { display: !isHomepage && `flex` },
-              [presets.Xl]: { display: `flex` },
+              [breakpoints.lg]: { display: `flex` },
             }}
           >
             <SocialNavItem href="https://gatsby.dev/discord" title="Discord">
@@ -156,7 +167,7 @@ const Navigation = ({ pathname }) => {
           <SocialNavItem
             href="https://www.gatsbyjs.com"
             title="gatsbyjs.com"
-            overrideCSS={{ paddingRight: 0 }}
+            overrideCSS={{ paddingRight: 0, fontSize: scale[2] }}
           >
             .com
           </SocialNavItem>
@@ -175,19 +186,21 @@ const styles = {
   },
   navContainer: {
     display: `none`,
-    [presets.Md]: {
+    [breakpoints.md]: {
       alignSelf: `flex-end`,
       display: `flex`,
+      marginLeft: space[6],
+      // marginLeft: isHomepage ? false : `auto`,
+      // marginRight: `auto`,
     },
   },
   ulContainer: {
     display: `none`,
-    [presets.Md]: {
+    [breakpoints.md]: {
       alignSelf: `flex-end`,
       display: `flex`,
       flexGrow: 1,
       margin: 0,
-      marginLeft: rhythm(1 / 4),
       listStyle: `none`,
       maskImage: `linear-gradient(to right, transparent, white ${rhythm(
         1 / 8
@@ -197,29 +210,26 @@ const styles = {
   },
   containerInner: {
     margin: `0 auto`,
-    paddingLeft: rhythm(3 / 4),
-    paddingRight: rhythm(3 / 4),
-    fontFamily: options.headerFontFamily.join(`,`),
+    paddingLeft: space[6],
+    paddingRight: space[6],
+    fontFamily: fonts.header,
     display: `flex`,
     alignItems: `center`,
     width: `100%`,
     height: `100%`,
+    position: `relative`,
   },
   navItem: {
-    ...scale(-1 / 3),
-    borderBottom: `0.125rem solid transparent`,
-    color: `inherit`,
+    fontSize: scale[3],
+    borderBottom: `2px solid transparent`,
+    color: colors.gray.calm,
     display: `block`,
-    letterSpacing: `0.03em`,
     WebkitFontSmoothing: `antialiased`,
-    lineHeight: `calc(${presets.headerHeight} - ${navItemTopOffset})`,
+    lineHeight: `calc(${dimensions.headerHeight} - ${navItemTopOffset})`,
     position: `relative`,
     textDecoration: `none`,
-    textTransform: `uppercase`,
     top: 0,
-    transition: `color ${presets.animation.speedDefault} ${
-      presets.animation.curveDefault
-    }`,
+    transition: `color ${transition.speed.default} ${transition.curve.default}`,
     zIndex: 1,
     "&:hover": {
       color: colors.gatsby,
@@ -240,18 +250,15 @@ const styles = {
     marginLeft: `auto`,
   },
   logo: {
-    height: 28,
+    height: space[6],
     margin: 0,
-    [presets.Md]: {
-      height: `1.55rem`,
-    },
   },
   logoLink: {
     alignItems: `center`,
     color: `inherit`,
     display: `flex`,
     flexShrink: 0,
-    marginRight: rhythm(1 / 2),
+    marginRight: space[3],
     textDecoration: `none`,
   },
 }