+++ www/src/components/package-readme.js
@@ -3,9 +3,11 @@ import PropTypes from "prop-types"
 import { Helmet } from "react-helmet"
 
 import { Link } from "gatsby"
+import { space } from "../utils/presets"
 import Container from "../components/container"
 import MarkdownPageFooter from "../components/markdown-page-footer"
 import GithubIcon from "react-icons/lib/go/mark-github"
+import GatsbyIcon from "../monogram.svg"
 import { linkStyles } from "../utils/styles"
 
 const PackageReadMe = props => {
@@ -26,24 +28,58 @@ const PackageReadMe = props => {
       <div
         css={{
           display: `flex`,
+          flexWrap: `wrap`,
           justifyContent: `space-between`,
+          paddingBottom: space[9],
+          "&&:hover": {
+            color: `inherit`,
+          },
         }}
       >
-        <a
-          css={{ ...linkStyles }}
-          href={githubUrl}
-          aria-labelledby="github-link-label"
+        <div
+          css={{
+            display: `flex`,
+            justifyContent: `space-between`,
+          }}
         >
-          <GithubIcon focusable="false" style={{ marginRight: `.5rem` }} />
-          <span id="github-link-label">View plugin on GitHub</span>
-        </a>
+          {githubUrl.indexOf(`https://github.com/gatsbyjs/gatsby`) === 0 &&
+            packageName[0] !== `@` && (
+              <div
+                css={{
+                  ...linkStyles,
+                  color: `#aaa !important`,
+                  marginRight: space[6],
+                }}
+              >
+                <img
+                  src={GatsbyIcon}
+                  css={{
+                    height: 16,
+                    marginBottom: 0,
+                    marginRight: 5,
+                    filter: `grayscale(100%)`,
+                    opacity: 0.5,
+                  }}
+                  alt={`Official Gatsby Plugin`}
+                />
+                Official Plugin
+              </div>
+            )}
+          <a
+            css={{ ...linkStyles }}
+            href={githubUrl}
+            aria-labelledby="github-link-label"
+          >
+            <GithubIcon focusable="false" style={{ marginRight: space[2] }} />
+            <span id="github-link-label">View plugin on GitHub</span>
+          </a>
+        </div>
         {githubUrl && (
           <Link to={`/starters?d=${packageName}`} css={{ ...linkStyles }}>
             See starters using this
           </Link>
         )}
       </div>
-
       <div
         css={{ position: `relative` }}
         dangerouslySetInnerHTML={{ __html: html }}